remove concept of extensible anies
Extensible any-types turns this into a degenerate
factory system, and that's what Any is trying to avoid.

You should not register global any types because it implies you have
a type that is global to every part of your system, which is bad design.

Instead, one can use modules to share custom any methods across instances.